Adjust test for new API

+++ src/Controls/tests/Xaml.UnitTests/MarkupExtensionTests.cs
@@ -46,9 +46,9 @@ namespace Microsoft.Maui.Controls.Xaml.UnitTests
 
 	public class ColorMarkup : IMarkupExtension
 	{
-		public int R { get; set; }
-		public int G { get; set; }
-		public int B { get; set; }
+		public byte R { get; set; }
+		public byte G { get; set; }
+		public byte B { get; set; }
 
 		public ColorMarkup()
 		{
